:recycle: Refactored components into seperated folders
