+++ music-app/src/components/dashboard/LikedSongs.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 /* import { Mixpanel } from '../analytics/Mixpanel'; */
 import { Grid, Typography } from '@material-ui/core';
 
-import { getlikedSongs, getUsers, getSpotifyAccountDetails } from '../actions';
+import { getlikedSongs, getUsers, getSpotifyAccountDetails } from '../../actions';
 import Song from './Song.js';
 
 class LikedSongs extends React.Component {
